Function Shift_Alpha
This is a function shift_alpha that takes in an alphabet string (A) and a shift constant (shift, default value of 0), prints the original and shifted alphabet (each as a single string), and returns a translation dictionary that contains each original letter as the key and the encrypted letter as the value. The performance of this function has been shown on the upper case alphabet with a shift of 0, +12, and -3.