Encrypt
This is a function encrypt that takes in a word and a shift (default value = 0), and returns a string containing the encrypted version of the word using a Caesar shift. Function can accept both upper and lower case letters, and returns any non-alphabetic characters as is. The use of this function has been demonstrated on the following list of (case-sensitive) words/phrases with a Caesar shift of +5: apple, Maryland, Route 66, LaTeX, twenty-first century